[RN] Don't override config values
