+++ react/features/app/functions.web.js
@@ -71,8 +71,6 @@ const _INTERCEPT_COMPONENT_RULES = [
     }
 ];
 
-export { getLocationContextRoot, _parseURIString } from './functions.native';
-
 /**
  * Determines which route is to be rendered in order to depict a specific Redux
  * store.
