+++ react/features/base/util/index.js
@@ -1,2 +1,3 @@
 export * from './loadScript';
 export * from './randomUtil';
+export * from './uri';
